further refinment of selectable design

+++ Codespirals.Base.Shared/Interfaces/EnforcementInterfaces/IOrderable.cs
@@ -2,6 +2,6 @@
 {
     public interface IOrderable
     {
-        public int Order { get; set; }
+        public short Order { get; set; }
     }
 }
