+++ Codespirals.Base.Shared/Interfaces/EnforcementInterfaces/IRankable.cs
@@ -2,6 +2,6 @@
 {
     public interface IRankable
     {
-        public int Rank { get; set; }
+        public short Rank { get; set; }
     }
 }
